Updated a lot!
Still waiting on parts so I've had a bash at tiding stuff up!

Major updates:

I've found a solotion that makes sure to only count one pulse from the lightsensor per shot, which means in theory 3-shot burst should be fixed!

Also, I've bought an active breaking mosfet so this should fix the overspinning but to fix the downsides of active breaking in full auto, I've added a function that makes sure to finish off the cycle/do one more after the trigger is released in full auto!

Finally, I've started to put together the structure for the final script with all the firemodes together, mainly putting everything into functions and figuring out some methods to manage the checking of the firemode.

Still need to come up with a way to switch the second fire mode between full auto and 3-shot burst. (usefull not just as a bonus but as active breaking can lead to lockups with the cutoff leaver left in, which I will be leaving in)